backend/requirements.txt
flask
pymongo
dnspython
python-dotenv == 1.2.1
flask-cors
jsonify